requirements.txt
python-dotenv==1.0.1
google_search_results==2.4.2
pandas==2.2.3